// src/app/enums/DutyTypesEnum.ts
export enum DutyTypesEnum {
    FixedDuty = '1160f693-bef5-e011-a485-80ee7300c611',
    FixedPeriodDuty = '1260f693-bef5-e011-a485-80ee7300c612',
    FreeDuty = '1360f693-bef5-e011-a485-80ee7300c613',
    Hourly = '1360f693-bef5-e011-a485-80ee7300c614',
    Daily = '1360f693-bef5-e011-a485-80ee7300c615'
}